requirements.txt
numpy
pandas
matplotlib
scikit-learn
scipy
tqdm
pydicom
scikit-image
tensorflow
umap-learn